Q6: appopriate typecasting

+++ q6/main.py
@@ -14,12 +14,12 @@ def findstrike(velocity: float, alt: float, az: float):
 	u = velocity
 	alt = math.radians(alt)
 	az = math.radians(az)
-	r1 = ((u**2)*(math.sin(2*az)))/g
-	r2 = ((u*math.cos(az))/g)
+	r1 = float(((u**2)*(math.sin(2*az)))/g)
+	r2 = float(((u*math.cos(az))/g))
 	r2_= math.sqrt(((u*math.sin(az))**2) + (2*g*h)) - (u*math.sin(az))
 	r2 = r2*r2_
 	r = r1+r2 	# arc distance
-	b = r/re 	# radial arc distance
+	b = float(r/re) 	# radial arc distance
 
 	a = math.acos(	math.cos(b)*math.cos(pi2-lat1) + (math.sin(pi2-lat1)*math.sin(b)*math.cos(az))	)
 	B = math.asin(	(math.sin(b)*math.sin(az))/math.sin(a)	)
